+++ tests/User/UserServiceTest.php
@@ -8,7 +8,6 @@ declare(strict_types = 1);
 namespace Wizaplace\SDK\Tests\User;
 
 use GuzzleHttp\Psr7\Uri;
-use InvalidArgumentException;
 use Wizaplace\SDK\Authentication\ApiKey;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Authentication\BadCredentials;
@@ -17,6 +16,7 @@ use Wizaplace\SDK\User\RegisterUserCommand;
 use Wizaplace\SDK\User\UpdateUserAddressCommand;
 use Wizaplace\SDK\User\UpdateUserAddressesCommand;
 use Wizaplace\SDK\User\UpdateUserCommand;
+use Wizaplace\SDK\User\UserAddress;
 use Wizaplace\SDK\User\UserAlreadyExists;
 use Wizaplace\SDK\User\UserService;
 use Wizaplace\SDK\User\UserTitle;
@@ -84,38 +84,36 @@ final class UserServiceTest extends ApiTestCase
         $userPassword = 'password';
         $userFistname = 'John';
         $userLastname = 'Doe';
-        $userAddresses = [
-            'billing'  => [
-                'title'     => UserTitle::MR(),
-                'firstname' => $userFistname,
-                'lastname'  => $userLastname,
-                'company'   => "Company_b",
-                'phone'     => "Phone_b",
-                'address'   => "Address_b",
-                'address_2' => "Address 2_b",
-                'zipcode'   => "Zipcode_b",
-                'city'      => "City_b",
-                'country'   => "FR",
-            ],
-            'shipping' => [
-                'title'     => UserTitle::MR(),
-                'firstname' => $userFistname,
-                'lastname'  => $userLastname,
-                'company'   => "Company_s",
-                'phone'     => "Phone_s",
-                'address'   => "Address_s",
-                'address_2' => "Address 2_s",
-                'zipcode'   => "Zipcode_s",
-                'city'      => "City_s",
-                'country'   => "FR",
-            ],
-        ];
+        $userBilling = new UserAddress([
+            'title'     => UserTitle::MR()->getValue(),
+            'firstname' => $userFistname,
+            'lastname'  => $userLastname,
+            'company'   => "Company_b",
+            'phone'     => "Phone_b",
+            'address'   => "Address_b",
+            'address_2' => "Address 2_b",
+            'zipcode'   => "Zipcode_b",
+            'city'      => "City_b",
+            'country'   => "FR",
+        ]);
+        $userShipping = new UserAddress([
+            'title'     => UserTitle::MR()->getValue(),
+            'firstname' => $userFistname,
+            'lastname'  => $userLastname,
+            'company'   => "Company_s",
+            'phone'     => "Phone_s",
+            'address'   => "Address_s",
+            'address_2' => "Address 2_s",
+            'zipcode'   => "Zipcode_s",
+            'city'      => "City_s",
+            'country'   => "FR",
+        ]);
 
         $client = $this->buildApiClient();
         $userService = new UserService($client);
 
         // create new user
-        $userId = $userService->register($userEmail, $userPassword, $userFistname, $userLastname, $userAddresses);
+        $userId = $userService->register($userEmail, $userPassword, $userFistname, $userLastname, $userBilling, $userShipping);
 
         // authenticate with newly created user
         $client->authenticate($userEmail, $userPassword);
@@ -134,85 +132,28 @@ final class UserServiceTest extends ApiTestCase
         $this->assertFalse($user->isVendor());
 
         // shipping address
-        $this->assertNull($user->getShippingAddress()->getTitle());
-        $this->assertSame($userFistname, $user->getShippingAddress()->getFirstName());
-        $this->assertSame($userLastname, $user->getShippingAddress()->getLastName());
-        $this->assertSame($userAddresses['shipping']['company'], $user->getShippingAddress()->getCompany());
-        $this->assertSame($userAddresses['shipping']['phone'], $user->getShippingAddress()->getPhone());
-        $this->assertSame($userAddresses['shipping']['address'], $user->getShippingAddress()->getAddress());
-        $this->assertSame($userAddresses['shipping']['address_2'], $user->getShippingAddress()->getAddressSecondLine());
-        $this->assertSame($userAddresses['shipping']['zipcode'], $user->getShippingAddress()->getZipCode());
-        $this->assertSame($userAddresses['shipping']['city'], $user->getShippingAddress()->getCity());
-        $this->assertSame('FR', $user->getShippingAddress()->getCountry());
+        $this->assertSame($userShipping->getTitle()->getValue(), $user->getShippingAddress()->getTitle()->getValue());
+        $this->assertSame($userShipping->getFirstName(), $user->getShippingAddress()->getFirstName());
+        $this->assertSame($userShipping->getLastName(), $user->getShippingAddress()->getLastName());
+        $this->assertSame($userShipping->getCompany(), $user->getShippingAddress()->getCompany());
+        $this->assertSame($userShipping->getPhone(), $user->getShippingAddress()->getPhone());
+        $this->assertSame($userShipping->getAddress(), $user->getShippingAddress()->getAddress());
+        $this->assertSame($userShipping->getAddressSecondLine(), $user->getShippingAddress()->getAddressSecondLine());
+        $this->assertSame($userShipping->getZipCode(), $user->getShippingAddress()->getZipCode());
+        $this->assertSame($userShipping->getCity(), $user->getShippingAddress()->getCity());
+        $this->assertSame($userShipping->getCountry(), $user->getShippingAddress()->getCountry());
 
         // billing address
-        $this->assertNull($user->getBillingAddress()->getTitle());
-        $this->assertSame($userFistname, $user->getBillingAddress()->getFirstName());
-        $this->assertSame($userLastname, $user->getBillingAddress()->getLastName());
-        $this->assertSame($userAddresses['billing']['company'], $user->getBillingAddress()->getCompany());
-        $this->assertSame($userAddresses['billing']['phone'], $user->getBillingAddress()->getPhone());
-        $this->assertSame($userAddresses['billing']['address'], $user->getBillingAddress()->getAddress());
-        $this->assertSame($userAddresses['billing']['address_2'], $user->getBillingAddress()->getAddressSecondLine());
-        $this->assertSame($userAddresses['billing']['zipcode'], $user->getBillingAddress()->getZipCode());
-        $this->assertSame($userAddresses['billing']['city'], $user->getBillingAddress()->getCity());
-        $this->assertSame('FR', $user->getBillingAddress()->getCountry());
-    }
-
-    public function testCreateUserWithAddressesWithMissingData()
-    {
-        $userEmail = 'user@example.com';
-        $userPassword = 'password';
-        $userFistname = 'John';
-        $userLastname = 'Doe';
-        $userAddresses = [
-            'shipping' => [
-                'title'     => UserTitle::MR(),
-                'firstname' => $userFistname,
-                'lastname'  => $userLastname,
-                'company'   => "Company_s",
-                'phone'     => "Phone_s",
-                'address'   => "Address_s",
-                'address_2' => "Address 2_s",
-                'zipcode'   => "Zipcode_s",
-                'city'      => "City_s",
-                'country'   => "FR",
-            ],
-        ];
-
-        $client = $this->buildApiClient();
-        $userService = new UserService($client);
-
-        // create new user
-        $this->expectException(InvalidArgumentException::class);
-        $userService->register($userEmail, $userPassword, $userFistname, $userLastname, $userAddresses);
-
-
-        $userAddresses = [
-            'billing'  => [
-                'title'     => UserTitle::MR(),
-                'firstname' => $userFistname,
-                'lastname'  => $userLastname,
-            ],
-            'shipping' => [
-                'title'     => UserTitle::MR(),
-                'firstname' => $userFistname,
-                'lastname'  => $userLastname,
-                'company'   => "Company_s",
-                'phone'     => "Phone_s",
-                'address'   => "Address_s",
-                'address_2' => "Address 2_s",
-                'zipcode'   => "Zipcode_s",
-                'city'      => "City_s",
-                'country'   => "FR",
-            ],
-        ];
-
-        $client = $this->buildApiClient();
-        $userService = new UserService($client);
-
-        // create new user
-        $this->expectException(InvalidArgumentException::class);
-        $userService->register($userEmail, $userPassword, $userFistname, $userLastname, $userAddresses);
+        $this->assertSame($userBilling->getTitle()->getValue(), $user->getBillingAddress()->getTitle()->getValue());
+        $this->assertSame($userBilling->getFirstName(), $user->getBillingAddress()->getFirstName());
+        $this->assertSame($userBilling->getLastName(), $user->getBillingAddress()->getLastName());
+        $this->assertSame($userBilling->getCompany(), $user->getBillingAddress()->getCompany());
+        $this->assertSame($userBilling->getPhone(), $user->getBillingAddress()->getPhone());
+        $this->assertSame($userBilling->getAddress(), $user->getBillingAddress()->getAddress());
+        $this->assertSame($userBilling->getAddressSecondLine(), $user->getBillingAddress()->getAddressSecondLine());
+        $this->assertSame($userBilling->getZipCode(), $user->getBillingAddress()->getZipCode());
+        $this->assertSame($userBilling->getCity(), $user->getBillingAddress()->getCity());
+        $this->assertSame($userBilling->getCountry(), $user->getBillingAddress()->getCountry());
     }
 
     public function testCreateUserWithFullInfos()