+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::addImageToMultivendorProduct`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::getGroupUsers`
-- Added optional parameter `$addresses` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
+- Added optionals parameters `$billing` and `$shipping` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
 
 ### Corrections